134 - Delete Todo with user ID

+++ TodoApp/routers/todos.py
@@ -84,11 +84,15 @@ async def update_todo(user : user_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code = status.HTTP_200_OK)
-async def delete_todo(db : db_dependency, todo_id : int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user : user_dependency,
+                      db : db_dependency, todo_id : int = Path(gt=0)):
+    todo_model = db.query(Todos)\
+                   .filter(Todos.id == todo_id)\
+                   .filter(Todos.owner_id == user.get('id'))\
+                   .first()
     if todo_model is None : 
         raise HTTPException(status_code = 404, detail = 'Todo not found.')
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
